% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_new_rows.R
\name{add_new_rows}
\alias{add_new_rows}
\title{Bind rows of two data frames if no duplication within specified column(s)}
\usage{
add_new_rows(df1, df2, ...)
}
\arguments{
\item{df1}{pre-existing data frame/tibble that you want to add rows to}

\item{df2}{new data frame/tibble that contain rows to add (must have exactly the same columns as df1)}

\item{...}{unquoted column names which specify grouping variables across which you want to check for duplication (no limit on how many or few you can use)}
}
\value{
If no duplication, a tibble containing unique rows from data frame 1 and 2. If duplication exists, halts execution with error message.
}
\description{
Bind rows of two data frames if no duplication within specified column(s)
}
\examples{
df1 <- dplyr::tibble(fin_year = c("2021/22", "2022/23"), kpi = "KPI 1.1", value = 1:2)
df2 <- dplyr::tibble(fin_year = "2023/24", kpi = "KPI 1.1", value = 2)

add_new_rows(df1, df2, fin_year, kpi)

#  A tibble: 3 × 3
#        fin_year     kpi    value
#          <chr>    <chr>   <dbl>
#   1     2021/22  KPI 1.1     1
#   2     2022/23  KPI 1.1     2
#   3     2023/24  KPI 1.1     2
}
