% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_new_rows_functions.R
\name{add_new_rows_tests}
\alias{add_new_rows_tests}
\title{Checks for add_new_rows() function}
\usage{
add_new_rows_tests(df1, df2)
}
\arguments{
\item{df1}{Dataframe, specified within add_new_rows()}

\item{df2}{Dataframe, specified within add_new_rows()}
}
\value{
Nothing, unless issue - then error
}
\description{
Checks for add_new_rows() function
}
