% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_history_checks.R
\name{build_history_checks}
\alias{build_history_checks}
\title{Check global environment for build_history() function.}
\usage{
build_history_checks(df_hist, df_new, kpi_number)
}
\arguments{
\item{df_hist}{Dataframe/tibble which contains historical data for relevant KPI.}

\item{df_new}{Dataframe/tibble containing new data for relevant KPI.}

\item{kpi_number}{KPI being added to the historical database, options are: "1.1-1.3", "1.4", "2", or "3"}
}
\value{
Nothing returned, just stops execution if error.
}
\description{
Ensures all variables required to run build_history() exist within the global environment.
}
\examples{
build_history_checks(old, new, "2")
}
