% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_history_functions.R
\name{build_history_create_new_hist}
\alias{build_history_create_new_hist}
\title{Helper to build_history() which creates new historical database and formats it}
\usage{
build_history_create_new_hist(
  df_hist,
  df_new,
  kpi_number,
  list_of_fys,
  list_of_hbs
)
}
\arguments{
\item{df_hist}{Dataframe/tibble which contains historical data for relevant KPI.}

\item{df_new}{Dataframe/tibble containing new data for relevant KPI.}

\item{kpi_number}{KPI being added to the historical database, options are: "1.1-1.3", "1.4", "2", or "3"}

\item{list_of_fys}{Character vector of financial years in data, in chronological order}

\item{list_of_hbs}{Character vector of Health Board names, in alphabetic order}
}
\value{
Creates new hist_db within build_history() environment
}
\description{
Helper to build_history() which creates new historical database and formats it
}
