% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_history_functions.R
\name{build_history_filenames}
\alias{build_history_filenames}
\title{Create filepaths for saving outputs of build_history() function.}
\usage{
build_history_filenames(kpi_number)
}
\arguments{
\item{kpi_number}{KPI being added to the historical database, options are: "1.1-1.3", "1.4", "2", or "3"}
}
\value{
List of variables, important ones are 1 filename_bckp and 2 filename_hist
}
\description{
Create filepaths for saving outputs of build_history() function.
}
