% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_history_format_df_new.R
\name{build_history_format_df_new}
\alias{build_history_format_df_new}
\title{Create and format new historical database for build_history()}
\usage{
build_history_format_df_new(kpi_number, df_new)
}
\arguments{
\item{kpi_number}{Passed from build_history()}

\item{df_new}{Passed from build_history()}
}
\value{
New historical database in the build_history() environment
}
\description{
Create and format new historical database for build_history()
}
