% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_aaa_gp_outputs.R
\name{create_aaa_gp_outputs}
\alias{create_aaa_gp_outputs}
\title{Create individual reports of GP coverage and self-referrals for Health Boards}
\usage{
create_aaa_gp_outputs(
  hb_name,
  financial_year,
  coverage_data,
  selfref_data,
  output_filepath,
  date_aaa_extracted,
  date_gp_extracted
)
}
\arguments{
\item{hb_name}{Character var of single Health Board}

\item{financial_year}{Character vector of financial year (long format 20XX/YY)}

\item{coverage_data}{Coverage data for current FY and previous FY, by GP practice}

\item{selfref_data}{Self-referrals data for current FY, by GP Practice}

\item{output_filepath}{Filepath to output folder (general)}

\item{date_aaa_extracted}{Date on which AAA extract was downloaded from Atos}

\item{date_gp_extracted}{Date on which GP history files were refreshed and downloaded from BOXI}
}
\value{
Saves Excel output for individual Health Board to 'GP Practices' folder at the specified output location.
}
\description{
Create individual reports of GP coverage and self-referrals for Health Boards
}
\examples{
create_aaa_gp_outputs(hb_name = "Borders",
                      financial_year = "2023/24",
                      coverage_data = gp_data,
                      selfref_data = sr_data,
                      output_filepath = "AAA/Topics/Screening/KPI/202409/output",
                      date_aaa_extracted = "1 September",
                      date_gp_extracted = "15 October")
}
