% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_seasonal_diff.R
\name{eval_seasonal_diff}
\alias{eval_seasonal_diff}
\title{Evaluate one of two expressions depending on the season.}
\usage{
eval_seasonal_diff(expr_spring, expr_autumn)
}
\arguments{
\item{expr_spring}{{expression} to be evaluated when season == "spring"}

\item{expr_autumn}{{expression} to be evaluated when season == "autumn"}
}
\value{
Runs function or assigns value to variable.
}
\description{
Function to evaluate spring- or autumn-specific expressions, cutting down on long ifelse statements.
Can be used to run functions alone, or assign values to variables in the global environment.
Note: function includes 'season' variable which should be defined as "spring" or "autumn" before using the function.
}
\examples{

season <- "autumn"

x <- eval_seasonal_diff({"it's spring!"}, {"it's autumn!"})

print(x)
[1] "it's autumn!"


}
