% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openxlsx_functions.R
\name{make_openxlsx_funcs}
\alias{make_openxlsx_funcs}
\title{Create custom openxlsx functions}
\usage{
make_openxlsx_funcs(wb_name, sheet_name)
}
\arguments{
\item{wb_name}{Unquoted name of 'workbook' active binding object}

\item{sheet_name}{Quoted name of worksheet to add to}
}
\value{
Wrappers for 5 openxlsx functions
}
\description{
Wrap-around functions for opexlsx's writeData, addStyle, mergeCells,
setColWidths, and setRowHeights that set the workbook name and sheet name
}
\examples{
make_openxlsx_funcs(wb, "Sheet1")
}
