% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{query_function}
\alias{query_function}
\title{Request confirmation for evaluating expression from user.}
\usage{
query_function(expression)
}
\arguments{
\item{expression}{An code snippet}
}
\value{
Evaluation of code snippet
}
\description{
Request confirmation for evaluating expression from user.
}
