% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{query_saveWorkbook}
\alias{query_saveWorkbook}
\title{Request confirmation for writing Excel workbooks from user.}
\usage{
query_saveWorkbook(wb, filepath)
}
\arguments{
\item{wb}{Workbook name that is being created using openxlsx package.}

\item{filepath}{Filepath to save output}
}
\value{
Prints confirmation of either writing or skipping.
}
\description{
Request confirmation for writing Excel workbooks from user.
}
