% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_write_rds.R
\name{query_write_rds}
\alias{query_write_rds}
\title{Request confirmation for writing rds objects from user.}
\usage{
query_write_rds(df, filepath)
}
\arguments{
\item{df}{A dataframe/tibble that's to be written out.}

\item{filepath}{The filepath to save the object to - must include filename.rds at end.}
}
\value{
Message indicating action which has been taken.
}
\description{
Request confirmation for writing rds objects from user.
}
\examples{

data <- data.frame(a = "1", b = "a")

query_write_rds(data, "temp/output_data.rds")
}
