% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_functions.R
\name{simplify_fy}
\alias{simplify_fy}
\title{Convert long-form financial year (20XX/YY) into short-form (XXYY)}
\usage{
simplify_fy(financial_year)
}
\arguments{
\item{financial_year}{Character string object/vector of financial year in format "20XX/YY"}
}
\value{
Character string object/vector of financial year in format "XXYY"
}
\description{
Convert long-form financial year (20XX/YY) into short-form (XXYY)
}
\examples{

simplify_fy("2023/24")
[1] "2324"

}
