% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_kpi_finyear.R
\name{viz_kpi_finyear}
\alias{viz_kpi_finyear}
\title{Create crosstab of 'kpi' and 'fin_year' for dataframe.}
\usage{
viz_kpi_finyear(df)
}
\arguments{
\item{df}{A dataframe/tibble containing columns 'kpi' and 'fin_year'}
}
\value{
Frequency table of 'kpi' x 'fin_year' for the specified dataframe.
}
\description{
Create crosstab of 'kpi' and 'fin_year' for dataframe.
}
