% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_history_create_new_hist.R
\name{build_history_create_new_hist}
\alias{build_history_create_new_hist}
\title{Helper to build_history() which creates new historical database and formats it}
\usage{
build_history_create_new_hist(df_hist, df_new, kpi_number)
}
\arguments{
\item{df_hist}{Passed from build_history()}

\item{df_new}{Passed from build_history()}

\item{kpi_number}{Passed from build_history()}
}
\value{
Creates new hist_db within build_history() environment
}
\description{
Helper to build_history() which creates new historical database and formats it
}
