% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_history_functions.R
\name{build_history_format_df_new}
\alias{build_history_format_df_new}
\title{Create and format new historical database for build_history()}
\usage{
build_history_format_df_new(kpi_number, df_new, fys_in_report)
}
\arguments{
\item{kpi_number}{KPI being added to the historical database, options are: "1.1-1.3", "1.4", "2", or "3"}

\item{df_new}{Dataframe/tibble containing new data for relevant KPI.}

\item{fys_in_report}{Character vector, must be >= 3 length. Includes financial year being published, plus two prior.}

\item{year_n}{Used for KPIs 1.1-1.3 or 1.4, character variable of a financial year. Default is 'year2', which corresponds to the financial year following that which is being published.}
}
\value{
New historical database in the build_history() environment
}
\description{
Create and format new historical database for build_history()
}
